Implement authorization using gates

+++ app/Http/Controllers/QuestionController.php
@@ -68,7 +68,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if(\Gate::allows('update-question', $question)){
+            return view('questions.edit', compact('question'));
+        }
+        abort(403, "Access denied");
     }
 
     /**
@@ -80,6 +83,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', 'Your question has been updated.');
     }
@@ -92,6 +98,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted.');
